Split out sys, sys msg, added init touch cmd

+++ bin/fake_rm.py
@@ -17,42 +17,18 @@ player_power = get_player_laser()
 player_attack_wait = .2
 
 
-
-def check_for_life(path):
-    with open(path) as fh:
-        test_line = fh.readline()
-        if test_line.startswith("life:"):
-            return(int(test_line.split(":")[-1].strip()))
-        else:
-            return(False)
-
-
-def get_respawn(path):
-    with open(path) as fh:
-        test_lines = fh.readlines()
-        for line in test_lines:
-            if line.startswith("respawn:"):
-                cut_line = ":".join(line.split(":")[1:])
-                return(cut_line)
-    
-    
-    
-def rewrite_life_file(path, value, respawn):
-    print(f"{path} Life:{value}")
-    data_to_write = f"life:{value}\nunder_attack:True\nrespawn:{respawn}"
-    with open(path, 'w') as fh:
-        fh.write(data_to_write)
-
-
 for thing_to_rm in args:
     if os.path.isfile(thing_to_rm):
         life = check_for_life(thing_to_rm)
         if life:
+            if friendly(thing_to_rm):
+                print(f"rm: cannot hurt ally {thing_to_rm}")
+                continue
             respawn = get_respawn(thing_to_rm)
-            print(f"found life: {thing_to_rm}")
+            #print(f"found life: {thing_to_rm}")
             for i in range(0,int(life/player_power)):
                 life = life - player_power
-                rewrite_life_file(thing_to_rm, life, respawn)
+                write_life_file(thing_to_rm, life, respawn, under_attack=True)
                 time.sleep(player_attack_wait)
             os.remove(thing_to_rm)
         else:
